敵スキルの親クラスを追加｜ * `Virus`スキルの継承を修正
