+++ MS_Project/Assets/Data/Character/WorldObjects.meta
