+++ MS_Project/Assets/Data/Character/WorldObjects/DataChicken.asset.meta
