+++ MS_Project/Assets/Data/Character/WorldObjects/DataVirus.asset
@@ -10,17 +10,17 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 488355e16bc0fbb4cbd0914b5457fdff, type: 3}
-  m_Name: DataChicken
+  m_Name: DataVirus
   m_EditorClassIdentifier: 
   maxHealth: 100
-  velocity: 1
-  fChaseDistance: 3
-  fDamage: 1
-  fAttackDistance: 0.8
-  fAttackCooldown: 1
-  enemyPrefab: Enemy/Chicken
-  eEnemyType: 1
-  onomatoData: {fileID: 11400000, guid: ab608adfc7c4b7842b9e89a7840a1f43, type: 2}
-  eTolerance: 3
+  gameObjPrefab: Enemy/Virus
+  SelfType: 2
+  onomatoData: {fileID: 11400000, guid: 8b0c05c99c42b564da4b9d77695334b1, type: 2}
+  tolerance: 1
   sfxClip: 
   vfxClip: 
+  actionCooldown: 1
+  damage: 1
+  attackDistance: 0.8
+  velocity: 1
+  chaseDistance: 3
