+++ MS_Project/Assets/Data/Character/WorldObjects/DataVirus.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8bac1e6146c45014fa164c534d078ad5
+guid: f2ae16b1b2db4ff4bb60f80b405a943e
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
