+++ MS_Project/Assets/Data/Character/WorldObjects/DataWalk.asset
@@ -1,0 +1,22 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cb2a201c6ebcc254db8f07456f46242d, type: 3}
+  m_Name: DataWalk
+  m_EditorClassIdentifier: 
+  maxHealth: 100
+  gameObjPrefab: 
+  SelfType: 5
+  onomatoData: {fileID: 11400000, guid: 8b0c05c99c42b564da4b9d77695334b1, type: 2}
+  tolerance: 0
+  sfxClip: 
+  vfxClip: 
+  actionCooldown: 1
