+++ MS_Project/Assets/Data/Character/WorldObjects/DataWalk.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 21bbf0cab483783479f755eadaa16c14
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
