+++ MS_Project/Assets/Resources/Enemy/EnemyContainer.prefab
@@ -68,10 +68,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e0ceae777d0280e45bb15ce11944ae87, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  player: {fileID: 0}
-  enemyObj: {fileID: 0}
-  onomatoObj: {fileID: 0}
-  status: {fileID: 0}
   fDamage: 0
   currentSpeed: 1
   isAttack: 1
@@ -82,21 +78,6 @@ MonoBehaviour:
       - m_Target: {fileID: 2106723936616505559}
         m_TargetAssemblyTypeName: 
         m_MethodName: set_MovementInput
-        m_Mode: 0
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: 
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
-  OnAttack:
-    m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 2106723936616505559}
-        m_TargetAssemblyTypeName: 
-        m_MethodName: Attack
         m_Mode: 0
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -121,6 +102,37 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
+  OnAttack:
+    m_PersistentCalls:
+      m_Calls:
+      - m_Target: {fileID: 2106723936616505559}
+        m_TargetAssemblyTypeName: 
+        m_MethodName: Attack
+        m_Mode: 0
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: 
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
+  OnUseSkill:
+    m_PersistentCalls:
+      m_Calls:
+      - m_Target: {fileID: 2106723936616505559}
+        m_TargetAssemblyTypeName: EnemyController, Assembly-CSharp
+        m_MethodName: UseSkill
+        m_Mode: 1
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
+  enemySkill: {fileID: 0}
 --- !u!54 &6712211598257689545
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -192,11 +204,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a91796c89bb97a042be996fac03c1139, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  statusData: {fileID: 11400000, guid: b861a65bec402a54eb23324324542341, type: 2}
+  statusData: {fileID: 0}
   currentHealth: 0
   moveSpeed: 0
   damage: 0
-  isDamaged: 0
-  enemyType: 0
+  selfType: 0
   onomatoData: {fileID: 0}
   tolerance: 0
