+++ MS_Project/Assets/Scenes/SampleScene.unity
@@ -732,7 +732,7 @@ PrefabInstance:
     - target: {fileID: 8981602233203678271, guid: d9b63e9ab6e49d94485e9a26fc2b4594, type: 3}
       propertyPath: statusData
       value: 
-      objectReference: {fileID: 11400000, guid: 8bac1e6146c45014fa164c534d078ad5, type: 2}
+      objectReference: {fileID: 11400000, guid: f2ae16b1b2db4ff4bb60f80b405a943e, type: 2}
     m_RemovedComponents: []
     m_RemovedGameObjects:
     - {fileID: 3910500974435302946, guid: d9b63e9ab6e49d94485e9a26fc2b4594, type: 3}
@@ -935,6 +935,10 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 8981602233203678271, guid: d9b63e9ab6e49d94485e9a26fc2b4594, type: 3}
+      propertyPath: statusData
+      value: 
+      objectReference: {fileID: 11400000, guid: b861a65bec402a54eb23324324542341, type: 2}
     m_RemovedComponents: []
     m_RemovedGameObjects:
     - {fileID: 3910500974435302946, guid: d9b63e9ab6e49d94485e9a26fc2b4594, type: 3}
