+++ MS_Project/Assets/Scripts/Character/WorldObjects/EnemySkills.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 41e13b97ffc1db84ea49143caf449265
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
