+++ MS_Project/Assets/Scripts/Character/WorldObjects/EnemySkills/EnemyHurt.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyHurt : EnemySkill
+{
+    public override void SkillAttack()
+    {
+        Debug.Log("UseSkill");
+    }
+}
