+++ MS_Project/Assets/Scripts/Character/WorldObjects/ObjectStatusManager.cs.meta
