+++ MS_Project/Assets/Scripts/Data/Character/CharacterStatusData.cs
@@ -5,9 +5,14 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "CharacterData", menuName = "ScriptableObjects/Character/CharacterStatusData", order = 0)]
 public class CharacterStatusData : ScriptableObject
 {
-    [Header("HP")]
+    [SerializeField, Header("オブジェクトタイプ")]
+    const WorldObjectType objectType = WorldObjectType.None;
+
+    [Header("最大HP")]
     public float maxHealth = 100.0f;
 
-    [Header("移動速度")]
-    public float velocity = 1.0f;
+    public WorldObjectType ObjectType
+    {
+        get => objectType;
+    }
 }
