+++ MS_Project/Assets/Scripts/Data/Character/WorldObjects/ObjectStatusData.cs
@@ -1,0 +1,31 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "ObjectData", menuName = "ScriptableObjects/WorldObjects/ObjectStatusData", order = 1)]
+public class ObjectStatusData : CharacterStatusData
+{
+    [SerializeField, Header("オブジェクトタイプ")]
+    const WorldObjectType objectType = WorldObjectType.StaticObject;
+
+    [Header("オブジェクトプレハブ")]
+    public string gameObjPrefab;
+
+    [Header("エネミータイプ")]
+    public OnomatoType SelfType = OnomatoType.None;
+
+    [Header("オノマトペデータ")]
+    public OnomatopoeiaData onomatoData;
+
+    [Header("耐性")]
+    public OnomatoType tolerance = OnomatoType.None;
+
+    [Header("SE")]
+    public string sfxClip;
+
+    [Header("VFX")]
+    public string vfxClip;
+
+    [Header("行動クールタイム")]
+    public float actionCooldown = 1f;
+}
