+++ MS_Project/Assets/Scripts/Data/Character/WorldObjects/ObjectStatusData.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cb2a201c6ebcc254db8f07456f46242d
